Moved caching implementation to separate file
